+++ libraries/datasource/src/main/java/androidx/media3/datasource/HttpEngineDataSource.java
@@ -759,8 +759,7 @@ public final class HttpEngineDataSource extends BaseDataSource implements HttpDa
     if (userAgent != null) {
       requestBuilder.addHeader(HttpHeaders.USER_AGENT, userAgent);
     }
-    // TODO: Uncomment when https://bugs.chromium.org/p/chromium/issues/detail?id=711810 is fixed
-    // (adjusting the code as necessary).
+    // TODO: crbug.com/41313195 - Uncomment when bug is fixed (adjusting the code as necessary).
     // Force identity encoding unless gzip is allowed.
     // if (!dataSpec.isFlagSet(DataSpec.FLAG_ALLOW_GZIP)) {
     //   requestBuilder.addHeader("Accept-Encoding", "identity");