+++ libraries/datasource/src/test/java/androidx/media3/datasource/UdpDataSourceContractTest.java
@@ -65,6 +65,7 @@ public class UdpDataSourceContractTest extends DataSourceContractTest {
             .setName("local-udp-unicast-socket")
             .setUri("udp://localhost:" + findFreeUdpPort())
             .setExpectedBytes(data)
+            .setMayResolveToUnknownLength(true)
             .build());
   }