+++ libraries/test_utils/src/test/java/androidx/media3/test/utils/FakeDataSourceContractTest.java
@@ -62,6 +62,7 @@ public class FakeDataSourceContractTest extends DataSourceContractTest {
             .setName("unknown length")
             .setUri(unknownLengthUri)
             .setExpectedBytes(unknownLengthData)
+            .setMayResolveToUnknownLength(true)
             .build());
   }